create Institution CRUD

+++ Modules/Core/Observers/CascadeSoftDeleteObserver.php
@@ -6,18 +6,20 @@ use Illuminate\Database\Eloquent\Model;
 
 class CascadeSoftDeleteObserver
 {
-    public function deleting(Model $model)
+    public function deleting(Model $model): void
     {
         if (property_exists($model, 'cascadeDeletes')) {
             foreach ($model->cascadeDeletes as $relation) {
                 foreach ($model->$relation as $related) {
-                    $related->delete();
+                    if (!is_null($related)) {
+                        $related->delete();
+                    }
                 }
             }
         }
     }
 
-    public function restoring(Model $model)
+    public function restoring(Model $model): void
     {
         if (request()->input('restore_cascade', 0) == 1) {
             if (property_exists($model, 'cascadeDeletes')) {
@@ -25,8 +27,10 @@ class CascadeSoftDeleteObserver
                     $model->$relation()
                             ->onlyTrashed()
                             ->get()
-                            ->each(function ($related) {
-                                $related->restore();
+                            ->each(function ($related): void {
+                                if (!is_null($related)) {
+                                    $related->restore();
+                                }
                             });
                 }
             }
